Refactor parsl tests
