+++ tests/parsl/test_syntax.py
@@ -5,10 +5,7 @@ import pytest
 parsl = pytest.importorskip("parsl")
 
 
-from pathlib import Path
-
-from quacc import SETTINGS, flow, job, redecorate, strip_decorator, subflow
-from quacc.wflow_tools.customizers import customize_funcs
+from quacc import flow, job, subflow
 
 
 def test_parsl_decorators(tmp_path, monkeypatch):
@@ -101,29 +98,6 @@ def test_parsl_decorators_args(tmp_path, monkeypatch):
     assert dynamic_workflow(1, 2, 3).result() == [6, 6, 6]
 
 
-def test_strip_decorators():
-    @job
-    def add(a, b):
-        return a + b
-
-    @flow
-    def add2(a, b):
-        return a + b
-
-    @subflow
-    def add3(a, b):
-        return a + b
-
-    stripped_add = strip_decorator(add)
-    assert stripped_add(1, 2) == 3
-
-    stripped_add2 = strip_decorator(add2)
-    assert stripped_add2(1, 2) == 3
-
-    stripped_add3 = strip_decorator(add3)
-    assert stripped_add3(1, 2) == 3
-
-
 def test_special_params(tmpdir, monkeypatch):
     monkeypatch.chdir(tmpdir)
 
@@ -138,39 +112,3 @@ def test_special_params(tmpdir, monkeypatch):
     assert add(1, 2).result() == 3
     assert add2(1, 2).result() == [4, 6, 8, 10, 12, 14]
 
-
-def test_change_settings_redecorate_job(tmp_path_factory):
-    tmp_dir1 = tmp_path_factory.mktemp("dir1")
-
-    @job
-    def write_file_job(name="job.txt"):
-        with open(Path(SETTINGS.RESULTS_DIR, name), "w") as f:
-            f.write("test file")
-
-    write_file_job = redecorate(
-        write_file_job, job(settings_swap={"RESULTS_DIR": tmp_dir1})
-    )
-    write_file_job().result()
-    assert Path(tmp_dir1 / "job.txt").exists()
-
-
-def test_change_settings_redecorate_flow(tmp_path_factory):
-    tmp_dir2 = tmp_path_factory.mktemp("dir2")
-
-    @job
-    def write_file_job(name="job.txt"):
-        with open(Path(SETTINGS.RESULTS_DIR, name), "w") as f:
-            f.write("test file")
-
-    @flow
-    def write_file_flow(name="flow.txt", job_decorators=None):
-        write_file_job_ = customize_funcs(
-            ["write_file_job"], [write_file_job], decorators=job_decorators
-        )
-        return write_file_job_(name=name)
-
-    # Test with redecorating a job in a flow
-    write_file_flow(
-        job_decorators={"write_file_job": job(settings_swap={"RESULTS_DIR": tmp_dir2})}
-    ).result()
-    assert Path(tmp_dir2 / "flow.txt").exists()
